create createApplication api

+++ src/main/java/com/sehs4701/repositiory/ScholarshipRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ScholarshipRepository extends JpaRepository<Scholarship, Long> {
+public interface ScholarshipRepository extends JpaRepository<Scholarship, Integer> {
 }
